feat: runner and separate spec cleanup
