+++ src/cala/models/__init__.py
@@ -1,4 +1,3 @@
 from .axis import AXIS
-from .spec import Pipe
 
-__all__ = [AXIS, Pipe]
+__all__ = [AXIS]
